pyproject.toml
[tool.poetry]
name = "custom-inst"
version = "0.1.0"
description = ""
authors = ["BogSyn <ravlykplus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
alembic = "^1.13.1"
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
uvicorn = {extras = ["standart"], version = "^0.29.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.6.4"}
python-multipart = "^0.0.9"
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
redis = "^5.0.3"
fastapi-limiter = "^0.1.6"
jinja2 = "^3.1.3"
cloudinary = "^1.39.1"
psycopg2 = "^2.9.9"
colorlog = "^6.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
